cloud_function/gcf_b64/requirements.txt
# Function dependencies, for example:
# package>=version
requests
tensorflow
efficientnet
google_api_python_client
protobuf
google-cloud-storage
